backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
SQLAlchemy==2.0.39
psycopg[binary]==3.2.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.8.1
python-multipart==0.0.20
